refactor: rename all deus packages to console packages
Rename all @repo/deus-* packages to @repo/console-* for better branding consistency.

Changes:
- Renamed packages:
  - deus-api-key → console-api-key
  - deus-api-services → console-api-services
  - deus-billing → console-billing
  - deus-octokit-github → console-octokit-github
  - deus-trpc → console-trpc
  - deus-types → console-types

- Updated package.json files:
  - Changed package names from @repo/deus-* to @repo/console-*
  - Updated internal dependencies to use new package names

- Updated apps/console/package.json dependencies
- Updated imports in console-api-services/src/base-service.ts
- Updated JSDoc comments in console-api-key/src/index.ts

Notes:
- Core deus package (core/deus) remains unchanged
- API package (api/deus) remains unchanged
- Database package (db/deus) remains unchanged
- Git history preserved using git mv
