+++ apps/console/package.json
@@ -26,10 +26,10 @@
 		"@hookform/resolvers": "catalog:",
 		"@octokit/openapi-types": "^26.0.0",
 		"@octokit/types": "^15.0.0",
-		"@repo/deus-api-services": "workspace:*",
-		"@repo/deus-octokit-github": "workspace:*",
-		"@repo/deus-trpc": "workspace:^",
-		"@repo/deus-types": "workspace:*",
+		"@repo/console-api-services": "workspace:*",
+		"@repo/console-octokit-github": "workspace:*",
+		"@repo/console-trpc": "workspace:^",
+		"@repo/console-types": "workspace:*",
 		"@repo/lib": "workspace:*",
 		"@repo/site-config": "workspace:*",
 		"@repo/ui": "workspace:*",
