+++ packages/console-api-services/src/errors.ts
