+++ packages/console-trpc/tsconfig.json
