+++ packages/console-types/src/session.ts
