--- packages/deus-types/package.json
@@ -1,51 +1,0 @@
-{
-  "name": "@repo/deus-types",
-  "version": "0.1.0",
-  "private": true,
-  "type": "module",
-  "exports": {
-    ".": {
-      "types": "./dist/index.d.ts",
-      "default": "./src/index.ts"
-    },
-    "./code-review": {
-      "types": "./dist/code-review.d.ts",
-      "default": "./src/code-review.ts"
-    },
-    "./repository": {
-      "types": "./dist/repository.d.ts",
-      "default": "./src/repository.ts"
-    },
-    "./session": {
-      "types": "./dist/session.d.ts",
-      "default": "./src/session.ts"
-    },
-    "./tasks": {
-      "types": "./dist/tasks.d.ts",
-      "default": "./src/tasks.ts"
-    }
-  },
-  "license": "MIT",
-  "scripts": {
-    "build": "tsc",
-    "clean": "git clean -xdf .cache .turbo dist node_modules",
-    "dev": "tsc --watch",
-    "format": "prettier --check . --ignore-path ../../.gitignore",
-    "lint": "eslint",
-    "typecheck": "tsc --noEmit --emitDeclarationOnly false"
-  },
-  "prettier": "@repo/prettier-config",
-  "dependencies": {
-    "ai": "catalog:",
-    "lightfast": "workspace:*"
-  },
-  "devDependencies": {
-    "@repo/eslint-config": "workspace:*",
-    "@repo/prettier-config": "workspace:*",
-    "@repo/typescript-config": "workspace:*",
-    "@types/node": "catalog:",
-    "eslint": "catalog:",
-    "prettier": "catalog:",
-    "typescript": "catalog:"
-  }
-}
